backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
chromadb==1.3.4
sentence-transformers==2.2.2
unstructured[local-inference]==0.18.20
python-docx==1.2.0
pypdf==6.3.0
openpyxl==3.1.5
requests==2.32.0
beautifulsoup4==4.13.5
google-generativeai==0.6.0
numpy==1.24.0
beautifulsoup4==4.13.5